Prevents generate-vendor.sh from failing when EXTRA_IMGS or OTA_IMGS is not defined.
In generate-vendor.sh, depending on the values of EXTRA_IMGS_LIST and
OTA_IMGS_LIST, those variable would be defined or not. There are checks to make
sure they are defined before using them, but such checks were missing at one
place. This commit fixes this.

This error was detected when preparing vendors for sailfish (Android 10):

    ./execute-all.sh -d sailfish -b QP1A.191005.007.A3 -o /path/to/AOSP

Resulting error:

    scripts/generate-vendor.sh: line 1016: EXTRA_IMGS[@]: unbound variable